Split out macOS and Linux base platform targets
So that we can have a separate mkdir function for each,
without replicating the code across x64 and arm64.
